+++ lib/atoms/icons/zwapIcons.dart
@@ -122,6 +122,8 @@ class ZwapIcons {
     'work': 'assets/images/icons/work.svg',
     'password': 'assets/images/icons/password.svg',
     "details": "assets/images/icons/Details.svg",
+    "open_eye": "assets/images/icons/open_eye.svg",
+    "closed_eye": "assets/images/icons/closed_eye.svg",
   };
 
   /// The mapping data for each icons