+++ lib/atoms/typography/zwapTypography.dart
@@ -35,6 +35,7 @@ enum ZwapTextType {
   smallBodyRegular,
   mediumBodyRegular,
   mediumBodyMedium,
+  mediumBodyBold,
   bigBodyRegular,
   bigBodySemibold,
 
@@ -274,7 +275,17 @@ class ZwapTypography {
   );
 
   static const TextStyle mediumBodyMedium = TextStyle(
-    fontWeight: FontWeight.w400,
+    fontWeight: FontWeight.w500,
+    fontStyle: FontStyle.normal,
+    fontSize: 14,
+    height: 1.43,
+    color: ZwapColors.shades100,
+    fontFamily: 'SFUIText',
+    package: 'zwap_design_system',
+  );
+
+  static const TextStyle mediumBodyBold = TextStyle(
+    fontWeight: FontWeight.w700,
     fontStyle: FontStyle.normal,
     fontSize: 14,
     height: 1.43,