+++ CHANGELOG.md
@@ -373,3 +373,9 @@
 
 ## 0.1.495
 * Added new text styles to ZwapTypography and ZwapText + textAlign propery in ZwapTextMultiStyle
+
+## 0.1.497
+* Animation in ZwapCheckbox and rounded icon instead of normal one
+* Added open eye and closed eye in ZwapIcons
+* Added custom prefix and suffix in ZwapInput
+* New ZwapTextMultiStyle with custom text styles and Link support